Refactored main.py
